fix: refactor project struct change project struct to RTSPserver and examples change how to build
