+++ README.md
@@ -30,5 +30,5 @@ sudo ./rtsp-opus-codec.out
 
 ### Check the sound by entering the speaker number in ffmpeg (Receive)
 ```bash
-ffmpeg -i rtsp://localhost:554 -f alsa default
+ffmpeg -i rtsp://localhost:8554 -f alsa default
 ```
