+++ example/CameraServer/FFmpegEncoder.h
