+++ src/RequestHandler.cpp
@@ -127,16 +127,17 @@ void RequestHandler::HandleDescribeRequest(const std::string& request, const int
     if (ParseAccept(request)) {
         response = "RTSP/1.0 200 OK\r\n";
 
-        if(ServerStream::getInstance().type == ServerStreamType::Audio) {
-            sdp = "v=0\r\n"
-              "o=- " + std::to_string(session->GetID()) + " " + std::to_string(session->GetVersion()) +
-              " IN IP4 " + ip + "\r\n"
-              "s=Opus Stream\r\n"
-              "c=IN IP4 " + ip + "\r\n"
-              "t=0 0\r\n"
-              "m=audio " + std::to_string(session->GetRTPPort()) + " RTP/AVP 111\r\n"  // Payload type for Opus
-              "a=rtpmap:111 opus/48000/2\r\n";  // Opus codec details
-        }else if(ServerStream::getInstance().type == ServerStreamType::Video) {
+        //TODO SDP 통합 필요
+        // if(ServerStream::getInstance().type == ServerStreamType::Audio) {
+        //     sdp = "v=0\r\n"
+        //       "o=- " + std::to_string(session->GetID()) + " " + std::to_string(session->GetVersion()) +
+        //       " IN IP4 " + ip + "\r\n"
+        //       "s=Opus Stream\r\n"
+        //       "c=IN IP4 " + ip + "\r\n"
+        //       "t=0 0\r\n"
+        //       "m=audio " + std::to_string(session->GetRTPPort()) + " RTP/AVP 111\r\n"  // Payload type for Opus
+        //       "a=rtpmap:111 opus/48000/2\r\n";  // Opus codec details
+        // }else if(ServerStream::getInstance().type == ServerStreamType::Video) {
             sdp = "v=0\r\n"
                 "o=- 0 0 IN IP4 " + ip + "\r\n"
                 "s=H264 Video Stream\r\n"
@@ -147,21 +148,7 @@ void RequestHandler::HandleDescribeRequest(const std::string& request, const int
                 "b=AS:40\r\n"
                 "a=rtpmap:96 H264/90000\r\n"
                 "a=fmtp:96 packetization-mode=1\r\n";
-        }
-
-        // }else if(ServerStream::getInstance().type == ServerStreamType::Video) {
-        //     sdp = "v=0\r\n"
-        //       "o=- " + std::to_string(session->GetID()) + " " + std::to_string(session->GetVersion()) +
-        //       " IN IP4 " + ip + "\r\n"
-        //       "s=H264 Video Stream\r\n"
-        //       "c=IN IP4 " + ip + "\r\n"
-        //       "t=0 0\r\n"
-        //       "m=video " + std::to_string(session->GetRTPPort()) + " RTP/AVP 96\r\n"
-        //       "a=rtpmap:96 H264/90000\r\n"
-        //       "a=control:track0\r\n";
-        // }
-        
-
+        //}
     } else {
         response = "RTSP/1.0 406 Not Acceptable\r\n";
     }
@@ -194,6 +181,8 @@ void RequestHandler::HandleSetupRequest(const std::string& request, const int cs
                            + "\r\n"
                              "\r\n";
     TCPHandler::GetInstance().SendRTSPResponse(session->GetTCPSocket(), response);
+
+    RTSPServer::getInstance().onInitEvent();  //TODO : play function when occur init event
 
     mediaStreamHandler = new MediaStreamHandler();
     mediaStreamHandler->udpHandler = new UDPHandler(session);
